+++ src/central_node_history.h
@@ -42,7 +42,7 @@ class History {
 
   int log(HistoryMessageType type, uint32_t id, uint32_t oldValue, uint32_t newValue, uint32_t aux);
 
-  int logFault(uint32_t id, uint32_t oldValue, uint32_t newValue, uint32_t faultStateId);
+  int logFault(uint32_t id, uint32_t oldValue, uint32_t newValue, uint32_t allowedClass);
   int logMitigation(uint32_t id, uint32_t oldValue, uint32_t newValue, uint32_t allowedClass);
   int logDeviceInput(uint32_t id, uint32_t oldValue, uint32_t newValue);
   int logAnalogDevice(uint32_t id, uint32_t oldValue, uint32_t newValue);